requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cycler==0.10.0
exceptiongroup==1.2.0
fastapi==0.103.2
fonttools==4.38.0
h11==0.14.0
httptools==0.6.0
idna==2.10
importlib-metadata==6.7.0
kiwisolver==1.4.5
matplotlib==3.5.3
numpy==1.21.6
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
packaging==23.2
Pillow==9.5.0
pydantic==2.5.3
pydantic-core==2.14.6
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.6
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
supervision==0.11.1
tqdm==4.66.1
typing-extensions==4.7.1
urllib3==2.0.7
uvicorn==0.22.0
uvloop==0.18.0
watchfiles==0.20.0
websockets==11.0.3
zipp==3.15.0
